Introduce a `PThreadExecutor`
# Motivation

We want to offer an executor that's based on `pthread`s and the platform's native selector system e.g. `epoll` or `kqueue`.

# Modifications

This PR lands the initial infrastructure to support a `pthread` based executor. This requires an internal C library and some abstractions around threads.

# Result

We have a working `pthread` based executor that we can use to build out a thread pool and add additional functionalities to it.