Appropriately size the default executor on Linux (#12)
# Motivation

The default executor on Linux should be appropriately sized to the
available cores of the system.

# Modifications

This PR configures the default Linux executor to use the available cores
of the system accounting for C1 and C2 group restrictions. Additionally,
it also allows to override the size by checking for an environment
variable.

# Result

Appropriate thread pool sizes on Linux while providing the user
customization options.