Implement `SchedulingExecutor` for `PThreadExecutor` on Linux (#19)
# Motivation

Our executors should be capable of scheduling sleeps for the continuous
and suspending clocks

# Modifications

This PR adds support to the `PThreadExecutor` when backed by the epoll
selector to handle sleeps for both the continuous and suspending clocks.
This is done by using the timer file descriptor support on Linux.

# Result

We can now handle clocks on our pthread based executor.